Refactor TentacleClient & script orchestration
