+++ source/Octopus.Tentacle.Client/Scripts/IScriptOrchestratorFactory.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Octopus.Tentacle.Client.Scripts
+{
+    interface IScriptOrchestratorFactory
+    {
+        Task<IScriptOrchestrator> CreateOrchestrator(CancellationToken cancellationToken);
+    }
+}
